Revert sync-status gating/caching/scheduler (#758, #761, #767) (#773)
Reverts:
- 2cb7b1fae fix: block data endpoints when indexer is syncing (#758)
- e95fb7482 feat: add 5s cache for sync status verification (#761)
- 990564bbc feat(sync) status refresh scheduler (#767)

These introduced verifySyncStatus() gating on /account/balance,
/account/coins, and /search, later backed by a periodic @Scheduled
background refresh sharing the Hikari connection pool with request
traffic. Under load, a transient DB timeout in the scheduled refresh
gets swallowed by PostgreSQLIndexCreationMonitor's defensive
exception handling and poisons the single shared cached status for
up to 10s, causing a step-function error spike across all gated
endpoints (e.g. 37.51% error rate on /account/balance at concurrency
24, 87.78% on /account/coins at concurrency 44).

This revert isolates the legacy stability test from that regression
to confirm the root cause before deciding on a permanent fix.
