+++ api/src/main/java/org/cardanofoundation/rosetta/api/account/controller/AccountApiImplementation.java
@@ -35,7 +35,6 @@ public class AccountApiImplementation implements AccountApi {
       throw ExceptionFactory.notSupportedInOfflineMode();
     }
     networkService.verifyNetworkRequest(accountBalanceRequest.getNetworkIdentifier());
-    networkService.verifySyncStatus();
 
     return ResponseEntity.ok(accountService.getAccountBalance(accountBalanceRequest));
   }
@@ -47,7 +46,6 @@ public class AccountApiImplementation implements AccountApi {
       throw ExceptionFactory.notSupportedInOfflineMode();
     }
     networkService.verifyNetworkRequest(accountCoinsRequest.getNetworkIdentifier());
-    networkService.verifySyncStatus();
 
     return ResponseEntity.ok(accountService.getAccountCoins(accountCoinsRequest));
   }
