+++ api/src/main/java/org/cardanofoundation/rosetta/api/network/service/NetworkService.java
@@ -20,6 +20,4 @@ public interface NetworkService {
 
     void verifyNetworkRequest(final NetworkIdentifier networkRequest);
 
-    void verifySyncStatus();
-
 }
