+++ api/src/main/java/org/cardanofoundation/rosetta/api/network/service/NetworkServiceImpl.java
@@ -25,7 +25,6 @@ import org.cardanofoundation.rosetta.api.block.service.LedgerBlockService;
 import org.cardanofoundation.rosetta.api.network.mapper.NetworkMapper;
 import org.cardanofoundation.rosetta.common.enumeration.OperationType;
 import org.cardanofoundation.rosetta.common.enumeration.OperationTypeStatus;
-import org.cardanofoundation.rosetta.common.enumeration.SyncStage;
 import org.cardanofoundation.rosetta.common.exception.ExceptionFactory;
 import org.cardanofoundation.rosetta.common.time.OfflineSlotService;
 import org.cardanofoundation.rosetta.common.util.Constants;
@@ -168,7 +167,7 @@ public class NetworkServiceImpl implements NetworkService {
       networkStatusBuilder.oldestBlock(oldestBlockIdentifier);
     }
 
-    syncStatusService.getSyncStatus().ifPresent(networkStatusBuilder::syncStatus);
+    syncStatusService.calculateSyncStatus(latestBlock).ifPresent(networkStatusBuilder::syncStatus);
 
     return networkStatusBuilder.build();
   }
@@ -182,15 +181,6 @@ public class NetworkServiceImpl implements NetworkService {
       if (!verifyNetwork(networkIdentifier.getNetwork())) {
         throw ExceptionFactory.networkNotFoundError();
       }
-    }
-  }
-
-  @Override
-  public void verifySyncStatus() {
-    SyncStatus syncStatus = syncStatusService.getSyncStatus()
-        .orElseThrow(ExceptionFactory::indexerNotReady);
-    if (!SyncStage.LIVE.getValue().equals(syncStatus.getStage())) {
-      throw ExceptionFactory.indexerNotReady();
     }
   }
