+++ api/src/main/java/org/cardanofoundation/rosetta/api/search/controller/SearchApiImpl.java
@@ -39,8 +39,8 @@ public class SearchApiImpl implements SearchApi {
     if (offlineMode) {
       throw ExceptionFactory.notSupportedInOfflineMode();
     }
+
     networkService.verifyNetworkRequest(searchTransactionsRequest.getNetworkIdentifier());
-    networkService.verifySyncStatus();
 
     Long limit = Optional.ofNullable(searchTransactionsRequest.getLimit())
             .orElse(LIMIT);
