+++ api/src/test/resources/testdata/errors.json
@@ -467,13 +467,6 @@
     "details": null
   },
   {
-    "code": 5056,
-    "message": "This endpoint is unavailable until the indexer data is ready. When /network/status returns \"stage\": \"LIVE\", this endpoint is ready for use.",
-    "description": null,
-    "retriable": true,
-    "details": null
-  },
-  {
     "code": 5058,
     "message": "Operation type filtering is not currently supported",
     "description": null,
