+++ tests/integration/golden_examples/rosetta_java/data/network/network_options_operation_types.json
@@ -497,11 +497,6 @@
           "retriable": false
         },
         {
-          "code": 5056,
-          "message": "This endpoint is unavailable until the indexer data is ready. When /network/status returns \"stage\": \"LIVE\", this endpoint is ready for use.",
-          "retriable": true
-        },
-        {
           "code": 5058,
           "message": "Operation type filtering is not currently supported",
           "retriable": false
